Allow for nrepl transport to be specified
nREPL supports a tty transport that allows for a user to telnet into nREPL
rather than having to use the default bencode protocol.

This is fairly useful with SublimeREPL as it doesn't have a 'connect
to a bencode repl' option for Clojure, it just has 'launch `lein
repl`' and 'telnet to a repl'.